backend/dbust-backend-fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.35.78
botocore==1.35.78
certifi==2024.12.14
click==8.1.7
dlib==19.24.6
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
jmespath==1.0.1
lazy_loader==0.4
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
s3transfer==0.10.4
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
sk-video==1.1.10
sniffio==1.3.1
starlette==0.41.3
tifffile==2024.9.20
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
